Update dependencies, adjust build to changes in 'metalsmith-layouts'
