+++ package.json
@@ -31,38 +31,41 @@
   "dependencies": {
     "autoprefixer-stylus": "0.14.0",
     "cheerio": "0.22.0",
-    "chokidar": "1.7.0",
-    "handlebars": "4.0.10",
+    "chokidar": "^2.0.3",
+    "handlebars": "^4.0.11",
+    "jstransformer-handlebars": "^1.1.0",
     "junk": "2.1.0",
     "lodash.defaultsdeep": "4.6.0",
-    "marked": "^0.3.9",
+    "marked": "^0.3.19",
     "metalsmith": "2.3.0",
     "metalsmith-collections": "0.9.0",
-    "metalsmith-feed": "0.5.0",
-    "metalsmith-layouts": "1.8.1",
-    "metalsmith-markdown": "0.2.1",
+    "metalsmith-discover-helpers": "^0.1.0",
+    "metalsmith-discover-partials": "^0.1.0",
+    "metalsmith-feed": "^1.0.0",
+    "metalsmith-layouts": "^2.1.0",
+    "metalsmith-markdown": "^0.2.2",
     "metalsmith-metadata": "0.0.4",
     "metalsmith-permalinks": "0.5.0",
-    "metalsmith-prism": "3.1.0",
+    "metalsmith-prism": "^3.1.1",
     "metalsmith-stylus": "3.0.0",
     "metalsmith-yearly-pagination": "2.0.0",
     "ncp": "2.0.0",
     "node-version-data": "1.0.1",
-    "octonode": "0.9.0",
-    "request": "2.83.0",
-    "require-dir": "0.3.2",
-    "semver": "5.4.1",
+    "octonode": "^0.9.2",
+    "request": "^2.85.0",
+    "require-dir": "^1.0.0",
+    "semver": "^5.5.0",
     "st": "1.2.2",
     "strftime": "0.10.0"
   },
   "devDependencies": {
     "faucet": "^0.0.1",
     "htmllint-cli": "^0.0.6",
-    "nock": "^9.0.2",
-    "pre-commit": "^1.1.2",
-    "proxyquire": "^1.7.4",
-    "standard": "^10.0.0",
-    "stylint": "^1.3.8",
-    "tape": "^4.5.1"
+    "nock": "^9.2.3",
+    "pre-commit": "^1.2.2",
+    "proxyquire": "^2.0.1",
+    "standard": "^11.0.1",
+    "stylint": "^1.5.9",
+    "tape": "^4.9.0"
   }
 }
