+++ scripts/helpers/json.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.exports = (context) => JSON.stringify(context)
